Rename (download -> storage) (future: upload too).
