+++ streaming/base/storage.py
